Fix minimap

+++ includes/cub3d.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   cub3d.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ebrodeur <ebrodeur@student.42lyon.fr>      +#+  +:+       +#+        */
+/*   By: wmonacho <wmonacho@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/01 14:26:25 by wmonacho          #+#    #+#             */
-/*   Updated: 2022/12/23 09:23:36 by ebrodeur         ###   ########lyon.fr   */
+/*   Updated: 2022/12/23 09:57:45 by wmonacho         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -112,8 +112,8 @@ typedef struct s_data {
 	int			free_tab_map;
 	int			textures_tab_size;
 	int			rgb_tab_size;
-	int			x_range;
-	int			y_range;
+	float		x_range;
+	float		y_range;
 	int			width;
 	char		*line_tmp;
 	char		*rgb_value;
